+++ service/meta/store/store.go
@@ -22,7 +22,7 @@ type Store interface {
 	Set(key, value string) error
 
 	// Set sets a value with key where the key exists or not
-	Incrby(key string, value int) error
+	Incrby(key string, value int64) error
 
 	// Delete deletes a key/value pair from the store
 	Delete(key string) error
